requirements.txt
pandas
pycryptodome